// src/types/core/Wallpaper.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Purity } from "./Purity";
import type { Thumbs } from "./Thumbs";

export type Wallpaper = { id: string, url: string, short_url: string, views: bigint, favorites: bigint, source: string, purity: Purity, category: string, dimension_x: number, dimension_y: number, resolution: string, ratio: string, file_size: bigint, file_type: string, created_at: string, colors: Array<string>, path: string, thumbs: Thumbs, };